Created an own component for DesktopMenu, moved menu components under Framework
