+++ src/IrcMonitor.ReactUi/src/framework/MenuBar.tsx
@@ -1,0 +1,67 @@
+import { AppBar, useMediaQuery, useTheme } from "@mui/material";
+import React, { useEffect } from "react";
+import { getChannels, getSelectecChannel, User } from "reducers/userReducer";
+
+import { useDispatch, useSelector } from "react-redux";
+import { userActions } from "actions/userActions";
+import { MobileMenu } from "./MobileMenu";
+import { DesktopMenu } from "./DesktopMenu";
+
+export interface MenuBarProps {
+  handleLoginWithGoogleAuthCode: () => void;
+  handleLogOut: () => void;
+  handleNavigateTo: (route: string) => void;
+  user: User | undefined;
+}
+
+const selectedChannelLocalStorageKey = "selectedChannel";
+
+export const MenuBar: React.FC<MenuBarProps> = ({
+  handleLogOut,
+  user,
+  handleNavigateTo,
+  handleLoginWithGoogleAuthCode
+}) => {
+  const theme = useTheme();
+  const drawDesktop = useMediaQuery(theme.breakpoints.up("lg"));
+
+  const channels = useSelector(getChannels);
+  const selectedChannel = useSelector(getSelectecChannel);
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    if (channels && channels.length > 0 && !selectedChannel) {
+      const valueInLocalStorage = localStorage.getItem(selectedChannelLocalStorageKey);
+
+      if (valueInLocalStorage && channels.some((c) => c.guid === valueInLocalStorage)) {
+        dispatch(userActions.selectChannel(valueInLocalStorage));
+      }
+    }
+  }, [channels, selectedChannel, dispatch]);
+
+  const handleSelectChannel = (channelId: string | undefined) => {
+    localStorage.setItem(selectedChannelLocalStorageKey, channelId);
+    dispatch(userActions.selectChannel(channelId));
+  };
+
+  return (
+    <AppBar position="fixed" sx={{ zIndex: (theme) => theme.zIndex.drawer + 1 }}>
+      {!drawDesktop ? (
+        <MobileMenu
+          onNavigate={handleNavigateTo}
+          onLogin={handleLoginWithGoogleAuthCode}
+          onLogOut={handleLogOut}
+          onSelectChannel={handleSelectChannel}
+        />
+      ) : (
+        <DesktopMenu
+          onNavigate={handleNavigateTo}
+          onLogin={handleLoginWithGoogleAuthCode}
+          onLogOut={handleLogOut}
+          onSelectChannel={handleSelectChannel}
+        />
+      )}
+    </AppBar>
+  );
+};
